@@ -10,18 +10,20 @@
     }
   ],
   "require": {
-    "php": "^7.2 || ^8.0",
-    "guzzlehttp/guzzle": "^6.5.5 || ^7.4.1",
-    "echron/tools": "^1.4.3",
+    "php": "^8.1",
+    "guzzlehttp/guzzle": "^6.5.8 || ^7.5.1",
+    "echron/tools": "dev-develop as 1.6.1",
     "echron/datatypes": "^1.4.1",
     "phpseclib/phpseclib": "^3.0.0",
-    "psr/simple-cache": "^1 || ^2 ||^3",
-    "league/flysystem-ftp": "^2.4.2"
+    "psr/simple-cache": "^3",
+    "psr/log": "^2.0 || ^3.0",
+    "league/flysystem-ftp": "^3.15.0"
   },
   "suggest": {
     "kunalvarma05/dropbox-php-sdk": "Use the Dropbox client",
     "google/apiclient": "Use the Google Drive client",
-    "aws/aws-sdk-php": "Use the AWS client"
+    "aws/aws-sdk-php": "Use the AWS client",
+    "symfony/cache": "Use Symfony cache clients"
   },
   "autoload": {
     "psr-4": {
@@ -30,10 +32,13 @@
   },
   "require-dev": {
     "phpunit/phpunit": "^9.0",
-    "cache/array-adapter": "^1.2.0",
     "kunalvarma05/dropbox-php-sdk": "^v0.4.1",
     "aws/aws-sdk-php": "^3.112.35",
-    "sempro/phpunit-pretty-print": "*"
+    "sempro/phpunit-pretty-print": "*",
+    "phpmd/phpmd": "^2.12.0",
+    "phpstan/phpstan": "^1.10.16",
+    "thecodingmachine/phpstan-safe-rule": "^v1.2.0",
+    "friendsofphp/php-cs-fixer": "^v3.17.0"
   },
   "require-disabled": {
     "google/apiclient": "^v2.12.6"
@@ -42,5 +47,10 @@
   "prefer-stable": true,
   "config": {
     "process-timeout": 0
+  },
+  "scripts": {
+    "lint:fix": " ./vendor/bin/php-cs-fixer fix --config=.php-cs-fixer.php",
+    "phpstan": " ./vendor/bin/phpstan analyse --memory-limit=512M",
+    "phpstan:generate-baseline": " ./vendor/bin/phpstan analyse --generate-baseline --memory-limit=512M"
   }
 }
