@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Echron\IO\Client;
@@ -9,13 +10,14 @@ use Echron\IO\Data\FileTransferInfo;
 use Exception;
 use GuzzleHttp\Client as GuzzleClient;
 use GuzzleHttp\ClientInterface;
+
 use function is_null;
 
 class Http extends Base
 {
-    private $guzzleClient;
+    private GuzzleClient $guzzleClient;
 
-    private $basicAuth;
+    private array|null $basicAuth = null;
 
     public function __construct(array $guzzleClientConfig = [])
     {
@@ -48,7 +50,7 @@ class Http extends Base
 
         $response = $this->guzzleClient->request('GET', $remote, $options);
         $contents = $response->getBody()
-                             ->getContents();
+            ->getContents();
         $this->setLocalFileContent($local, $contents);
 
         if (!is_null($localChangeDate)) {
